% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inhaler.R
\name{run_inhaler}
\alias{run_inhaler}
\title{Run inhaler analysis with new data}
\usage{
run_inhaler(newdata, prior = NULL)
}
\arguments{
\item{newdata}{New data set of the same structure as \code{inhaler} 
in `brms`}

\item{prior}{Optional new prior}
}
\value{
A structure of class brmsfit
}
\description{
Runs the inhaler example from brms/brm with new data set.
It is assumed that a a structure brmsfit has been generated by 
\code{brms/generate_stan.R}.
}
\examples{
run_inhaler()
}
