% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inhaler.R
\name{run_inhaler}
\alias{run_inhaler}
\title{Run inhaler analysis with new data}
\usage{
run_inhaler(newdata)
}
\arguments{
\item{newdata}{New data set of the same structure as \code{inhaler} 
in `brms`}
}
\value{
A structure of class brmsfit
}
\description{
Runs the inhaler example from brms/brm with new data set.
It is assumed that a a structure brmsfit has been generated by 
\code{brms/generate_stan.R}.
}
\examples{
# This example fails during build but works after installation
\dontrun{
data("inhaler", package= "brms")
run_inhaler(inhaler)
}
}
